Break out TestDataGenerator from generate_per_layer_test.py
TestDataGenerator is also used by gen_micro_mutable_op_resolver_test.py.
TestDataGenerator.generate_test() and generate_build_file() are moved
out. They have to be implemented by the inheriting class
since they are not generic and would be different for every user.
Adds generate_golden() to TestDataGenerator since it is generic.
