+++ tensorflow/lite/micro/integration_tests/BUILD
@@ -24,5 +24,6 @@ py_binary(
         "//tensorflow/lite/python:schema_py",
         "//tensorflow/lite/python:schema_util",
         "//tensorflow/lite/tools:flatbuffer_utils",
+        "//tensorflow/lite/micro/tools:generate_test_for_model",
     ],
 )
