+++ tensorflow/lite/micro/integration_tests/generate_per_layer_tests.py
@@ -29,6 +29,8 @@ from tensorflow.python.platform import gfile
 from tflite_micro.tensorflow.lite.python import schema_py_generated as schema_fb
 from tflite_micro.tensorflow.lite.python import schema_util
 from tflite_micro.tensorflow.lite.tools import flatbuffer_utils
+from tflite_micro.tensorflow.lite.micro.tools import generate_test_for_model
+
 
 TEMPLATE_DIR = os.path.join(os.path.dirname(__file__), 'templates')
 TEMPLATE_DIR = os.path.abspath(TEMPLATE_DIR)
@@ -136,150 +138,7 @@ class TestModelGenerator:
     return output_models
 
 
-class TestDataGenerator:
-
-  def __init__(self, output_dir, model_paths, inputs):
-    self.output_dir = output_dir
-    self.model_paths = model_paths
-    self.csv_filenames = []
-    self.inputs = inputs
-    self.input_types = {}
-    self.cc_srcs = []
-    self.cc_hdrs = []
-    self.includes = []
-
-  def generate_inputs_single(self, interpreter, dtype):
-    input_tensor = interpreter.tensor(
-        interpreter.get_input_details()[0]['index'])
-    return [
-        np.random.randint(low=np.iinfo(dtype).min,
-                          high=np.iinfo(dtype).max,
-                          dtype=dtype,
-                          size=input_tensor().shape),
-    ]
-
-  def generate_inputs_add_sub(self, interpreter, dtype):
-    input_tensor0 = interpreter.tensor(
-        interpreter.get_input_details()[0]['index'])
-    input_tensor1 = interpreter.tensor(
-        interpreter.get_input_details()[1]['index'])
-    return [
-        np.random.randint(low=np.iinfo(dtype).min,
-                          high=np.iinfo(dtype).max,
-                          dtype=dtype,
-                          size=input_tensor0().shape),
-        np.random.randint(low=np.iinfo(dtype).min,
-                          high=np.iinfo(dtype).max,
-                          dtype=dtype,
-                          size=input_tensor1().shape)
-    ]
-
-  def generate_inputs_transpose_conv(self, interpreter, dtype):
-    input_tensor0 = interpreter.tensor(0)
-    filter_tensor = interpreter.tensor(1)
-    input_tensor1 = interpreter.tensor(2)
-    output_shape = np.array([
-        1,
-        filter_tensor().shape[1] + input_tensor1().shape[1] - 1,
-        int(input_tensor1().shape[2] * 2 + filter_tensor().shape[2] / 2),
-        filter_tensor().shape[0]
-    ],
-                            dtype=np.int32)
-    return [
-        output_shape,
-        np.random.randint(low=np.iinfo(dtype).min,
-                          high=np.iinfo(dtype).max,
-                          dtype=dtype,
-                          size=input_tensor1().shape)
-    ]
-
-  def GetTypeStringFromTensor(self, tensor):
-    if tensor.dtype == np.int8:
-      return 'int8'
-    if tensor.dtype == np.int16:
-      return 'int16'
-    if tensor.dtype == np.int32:
-      return 'int32'
-    if tensor.dtype == np.float:
-      return 'float'
-
-  def generate_goldens(self, builtin_operator):
-    for model_path in self.model_paths:
-      # Load model and run a single inference with random inputs.
-      interpreter = tf.lite.Interpreter(model_path=model_path)
-      interpreter.allocate_tensors()
-      input_tensor = interpreter.tensor(
-          interpreter.get_input_details()[0]['index'])
-      output_tensor = interpreter.tensor(
-          interpreter.get_output_details()[0]['index'])
-
-      if builtin_operator in (schema_fb.BuiltinOperator.CONV_2D,
-                              schema_fb.BuiltinOperator.DEPTHWISE_CONV_2D,
-                              schema_fb.BuiltinOperator.STRIDED_SLICE,
-                              schema_fb.BuiltinOperator.PAD,
-                              schema_fb.BuiltinOperator.LEAKY_RELU):
-        generated_inputs = self.generate_inputs_single(interpreter,
-                                                       input_tensor().dtype)
-      elif builtin_operator in (schema_fb.BuiltinOperator.ADD,
-                                schema_fb.BuiltinOperator.SUB):
-        generated_inputs = self.generate_inputs_add_sub(
-            interpreter,
-            input_tensor().dtype)
-      elif builtin_operator == schema_fb.BuiltinOperator.TRANSPOSE_CONV:
-        input_tensor = interpreter.tensor(
-            interpreter.get_input_details()[1]['index'])
-        generated_inputs = self.generate_inputs_transpose_conv(
-            interpreter,
-            input_tensor().dtype)
-      else:
-        raise RuntimeError(f'Unsupported BuiltinOperator: {builtin_operator}')
-
-      for idx, input_tensor_idx in enumerate(self.inputs):
-        interpreter.set_tensor(input_tensor_idx, generated_inputs[idx])
-      interpreter.invoke()
-
-      self.write_golden(generated_inputs, model_path, output_tensor)
-
-  def write_golden(self, generated_inputs, model_path, output_tensor):
-    for input_idx, input_tensor_data in enumerate(generated_inputs):
-      input_type = self.GetTypeStringFromTensor(input_tensor_data)
-      self.input_types[input_idx] = input_type
-      input_flat = input_tensor_data.flatten().tolist()
-      csv_input_filename = \
-          f"{model_path.split('.')[0]}_input{input_idx}_{input_type}.csv"
-      input_csvfile = open(csv_input_filename, 'w', newline='')
-      input_csvwriter = csv.writer(input_csvfile)
-      input_csvwriter.writerow(input_flat)
-      self.csv_filenames.append(csv_input_filename)
-
-    output_flat = output_tensor().flatten().tolist()
-
-    # Write inputs and goldens to CSV file.
-    output_type = self.GetTypeStringFromTensor(output_tensor())
-    self.output_type = output_type
-    csv_golden_filename = f"{model_path.split('.')[0]}_golden_{output_type}.csv"
-    golden_csvfile = open(csv_golden_filename, 'w', newline='')
-    golden_csvwriter = csv.writer(golden_csvfile)
-    np.set_printoptions(threshold=np.inf)
-    golden_csvwriter.writerow(output_flat)
-    self.csv_filenames.append(csv_golden_filename)
-
-    # Collect all target names into a list
-    targets = []
-    for model_path in self.model_paths:
-      target_name = model_path.split('/')[-1].split('.')[0]
-      targets.append(target_name)
-
-    template_file_path = os.path.join(TEMPLATE_DIR, 'BUILD.mako')
-    build_template = template.Template(filename=template_file_path)
-    with open(self.output_dir + '/BUILD', 'w') as file_obj:
-      key_values_in_template = {
-          'targets': targets,
-          'inputs': self.inputs,
-          'input_dtypes': self.input_types,
-          'output_dtype': self.output_type
-      }
-      file_obj.write(build_template.render(**key_values_in_template))
+class TestPerLayerDataGenerator(generate_test_for_model.TestDataGenerator):
 
   def generate_tests(self):
     # Collect all target names into a list
@@ -304,27 +163,23 @@ class TestDataGenerator:
       }
       file_obj.write(build_template.render(**key_values_in_template))
 
-  def generate_makefile(self, test_file='integration_tests.cc', src_prefix=None):
-    makefile = open(self.output_dir + '/Makefile.inc', 'w')
-    output_dir_list = self.output_dir.split('/')
-    if src_prefix is None:
-      src_prefix = output_dir_list[-3] + '_' + output_dir_list[
-          -2] + '_' + output_dir_list[-1]
-    makefile.write(src_prefix + '_GENERATOR_INPUTS := \\\n')
+  def generate_build_file(self):
+    # Collect all target names into a list
+    targets = []
     for model_path in self.model_paths:
-      makefile.write(
-          model_path.split('third_party/tflite_micro/')[-1] + ' \\\n')
-    for csv_input in self.csv_filenames:
-      makefile.write(
-          csv_input.split('third_party/tflite_micro/')[-1] + ' \\\n')
-    makefile.write('\n')
-    makefile.write(src_prefix + '_SRCS := \\\n')
-    makefile.write(self.output_dir.split('third_party/tflite_micro/')[-1] +
-                   '/' + test_file)
-    makefile.write('\n\n')
-    makefile.write('$(eval $(call microlite_test,' + src_prefix + '_test,\\\n')
-    makefile.write('$(' + src_prefix + '_SRCS),,$(' + src_prefix +
-                   '_GENERATOR_INPUTS)))')
+      target_name = model_path.split('/')[-1].split('.')[0]
+      targets.append(target_name)
+
+    template_file_path = os.path.join(TEMPLATE_DIR, 'BUILD.mako')
+    build_template = template.Template(filename=template_file_path)
+    with open(self.output_dir + '/BUILD', 'w') as file_obj:
+      key_values_in_template = {
+          'targets': targets,
+          'inputs': self.inputs,
+          'input_dtypes': self.input_types,
+          'output_dtype': self.output_type
+      }
+      file_obj.write(build_template.render(**key_values_in_template))
 
 
 def op_info_from_name(name):
@@ -364,7 +219,7 @@ def main(_):
   inputs, builtin_operator = op_info_from_name(FLAGS.output_dir.split('/')[-1])
   generator = TestModelGenerator(model, FLAGS.output_dir, inputs)
   model_names = generator.generate_models(0, builtin_operator)
-  data_generator = TestDataGenerator(FLAGS.output_dir, model_names, inputs)
+  data_generator = TestPerLayerDataGenerator(FLAGS.output_dir, model_names, inputs)
   data_generator.generate_goldens(builtin_operator)
   data_generator.generate_build_file()
   data_generator.generate_makefile()
