+++ tensorflow/lite/micro/tools/BUILD
@@ -1,3 +1,14 @@
+py_library(
+    name = "generate_test_for_model",
+    srcs = ["generate_test_for_model.py"],
+    srcs_version = "PY3",
+    visibility = ["//:__subpackages__"],
+    deps = [
+        "//tensorflow/lite/python:schema_py",
+    ],
+
+)
+
 package(
     default_visibility = ["//:__subpackages__"],
     licenses = ["notice"],
